Add waitForMakers function and avoid segfaults when writing to file.
Use this to wait for any writing of markers to finish. Used when the dumping to file is stopped,
and when the I/O thread is stopped. Sometimes a marker was written after
closing the dump file. Check if the dumpfile is not nullptr both when
writing a marker and when writing power data